feat: post 관련 모든 service 추가 '

+++ src/main/java/com/example/bojeung365api/repository/scam/ScamReportPostRepository.java
@@ -3,5 +3,5 @@ package com.example.bojeung365api.repository.scam;
 import com.example.bojeung365api.entity.post.ScamReportPost;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ScamReportPostRepository extends JpaRepository<ScamReportPost, Integer>, ScamReportPostRepositoryCustom {
+public interface ScamReportPostRepository extends JpaRepository<ScamReportPost, Long>, ScamReportPostRepositoryCustom {
 }
